requirements.txt
Django==2.0.6
django-crispy-forms==1.7.2
Pillow==5.2.0
pytz==2018.5



